+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API.hs
@@ -4,6 +4,7 @@ module Cardano.Ledger.Shelley.API (
 ) where
 
 import Cardano.Ledger.Shelley.API.ByronTranslation as X
+import Cardano.Ledger.Shelley.API.Forecast as X
 import Cardano.Ledger.Shelley.API.Mempool as X
 import Cardano.Ledger.Shelley.API.Types as X
 import Cardano.Ledger.Shelley.API.Validation as X